Add files via upload
This is an example of what happens when the ui struggles to handle the number of inputs. The input for this drawing was a simple smile. The code seems to particularly struggle with handling the smile due to the curved nature of the line and the row by row coloring design. This leads to a possible future in identifying and drawing the unique strokes, whether it is curved or straight rather than dismissing the content and going line by line.